refactor: remove cycle from DocumentSigner
